Store decks in an SQLite database
The database is created automatically if it doesn't exist.
Quizzes and questions are stored in separate tables with a foreign
key in the questions table pointing to a quiz.

Right now, the QuizLoaderFile class isn't doing anything. Eventually
there will be an import/export system which will use the class
to parse deck files before saving to the database and export them.

Because of the SQLite dll, the project is set to only target the
x64 architecture.